Did some changes

+++ modules/db.js
@@ -24,9 +24,37 @@ dbMethods.createBlogPost = function(heading, blogtext, userid) {
 }
 
 // ------------------------------------
-dbMethods.deleteBlogPost = function(id) {  
-    let sql = "DELETE FROM blogposts WHERE id = $1 RETURNING *";
-	let values = [id];	
+dbMethods.deleteBlogPost = function(id, userid) {  
+    let sql = "DELETE FROM blogposts WHERE id = $1 AND userid = $2 RETURNING *";
+	let values = [id, userid];	
+    return pool.query(sql, values); //return the promise
+}
+
+// ------------------------------------
+dbMethods.getAllUsers = function() {
+    let sql = "SELECT id, username FROM users";
+    return pool.query(sql); //return the promise
+}
+
+// ------------------------------------
+dbMethods.getUser = function(username) {
+    let sql = "SELECT * FROM users WHERE username = $1";
+    let values = [username];
+    return pool.query(sql, values); //return the promise
+}
+
+// ------------------------------------
+dbMethods.createUser = function(username, password, salt) {
+    let sql = "INSERT INTO users (id, username, password, salt) VALUES(DEFAULT, $1, $2, $3) returning *";
+    let values = [username, password, salt];
+    //console.log(salt);
+    return pool.query(sql, values); //return the promise
+}
+
+// ------------------------------------
+dbMethods.deleteUser = function(id) {
+    let sql = "DELETE FROM users WHERE id = $1 RETURNING *";
+    let values = [id];
     return pool.query(sql, values); //return the promise
 }
